+++ src/utils/constants.ts
@@ -1,1 +1,11 @@
 export const APP_VERSION = 0.2
+
+export const FETCH_METHOD = {
+  GET_FULL_LIST: 'getFullList',
+}
+
+export const COLLECTIONS = {
+  USERS: 'users',
+  ARTICLES: 'articles',
+  TEST: 'test',
+}