+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Version 0.2
 
-- Implement PocketBase (Include first test in Dashboard.tsx)
+- Implement PocketBase (Include first test in adminDashboard.tsx)
 - Add Dashboard and login page for administrators
 - Add authentification context for keeping the connexion on the website